Path searching improved; path history calculation implemented.
Lots of changes; use of records, iterators, due to a better
understanding of how Chapel works (although by no means perfect).
Still a lot to learn, and I'm sure I can improve on what I'm doing.  I
have a feeling that I'm reinventing the wheel in a lot of places, but at
least for a proof of concept a lot of that can be adjusted later without
too much hassle, I'm sure.