Fixed deadlock and sync issues.
Read/writes work properly!  However, we're a bit slow; likely because
the algorithm I'm using to farm out tasks is pretty inefficient.
Essentially, each worker has to calculate a path to each node we're
going to process, rather than only doing it for the ones that it's
actually going to process.

I need to find a way to skip the path search if the node has already
been processed.  Otherwise, it's all just silly overhead.