Ha ha ha!  Fixed!
That is, no more deadlocking, no more ending the path search algorithm
too early (I was catching whether the currentNode that we're visiting in
djikstra's algo was part of the potential target nodes, id_B.  This is
fine, _except_ for when id_B is empty, which happens when the generation
ends.  In that event, the loop happened once, determined that there was
nowehere else to go, then ended in a break command to avoid an infinite
loop.  This is because the initial implementation did not always have a
target in mind.  At this point, whatever the last node we visited is the
node that we would end up processing.  This had the effect of
overwriting nodes we've already processed, destroying the set of edges
(and indeed, totally changing the seeds) and also inappropriately adding
in older nodes to the NEXT generation, resulting in a cascading failure.

Now, we catch whether we're empty and fail out, and our generations seem
to work just fine.  Plus, we seem to have linear scaling if you have N
valkyries for N seeds (ultimately this won't always be the case, given
that we'll have merge events, but).