Fixed domain clearing errors.
Switching happens quickly, now, and we can also create shortcuts between
nodes if the distance we're traveling is greater than a certain amount.
That should help with node traversal (and indeed, in tests, sped it up
by over an order of magnitude, particularly as the length of the graph
increased).