Implemented priority queue and 'move anyway'.
Nodes which a particular valkyrie creates are now added to a priority
queue.  That is, it tries to move to those nodes first, and worries
about the ones that need to be processed later.  This results in
dramatic speed efficiencies.  In addition, in the event that Valkyries
are not used, they'll still make a move; if they stay at root, then
processing becomes quite slow, as their search time to reach the current
generation increases substantially the further behind they get.

They don't _process_ anything, mind you; they just move to keep
themselves in sync.  Brings us back to linear scaling.