Massively overhauled the logging system.
Now each valkyrie prints to its own log; much more convenient.  In
addition, by default, we send normal, non-Valkyrie output to stdout, but
we also go ahead and save that in a logfile, too.  Makes it much easier
to debug, and should improve the speed of production once we switch to a
multilocale setting.

There are still a few edge bugs which pop up randomly (odd infinite
loops), and there are certainly places we could improve the speed,
but for now this is probably sufficient.