I do think our merge functions are actually working.
There are problems with creating unit tests for this sort of thing,
however, as eventually you run into overlap.  In addition, we're
triggering a lot of critical warnings during the unit test that are
false positives; I should determine whether the two numbers are within a
certain tolerance to make up for the fact that we're using ints and
reals.