Unfixable deadlocks.
HEY!  Not as bad as it sounds!  In the event that you try and use more
threads than you have available, I think the OS or chapel tries to
switch between tasks when some sort of criteria is met.  In the event
that you have handles left open on a particular task, you can become
deadlocked waiting for a write.

Assuming that's the case (I genuinely can't figure out what else it
might be, as I'm only deadlocking when I try and use more tasks than I
have cores), it seems a bit silly that it doesn't try and do work on the
others.  Is that a bug?  I can certainly seem to work around it by
limiting the number of concurrent tasks, but there's no reason why we
shouldn't be able to run it that way.