Increased main loop efficiency.
__calculatePath__ now has an option to allow it to check and set an
atomic; essentially allowing us to continue searching for the nearest
unprocessed node if some of the nodes that were considered potentials
have already been taken.  While the race condition didn't result in
errors, it did result in a lot of wasted computation as we lost our
results and had to recalculate without the old target.

Now, we do the testAndSet in __calculatePath__, and then just return if
it was unprocessed.