Oh what a tangled web and all that.
Need to restructure the C bits, unsurprisingly; we work super well in a
non-threaded environment.  Too bad we're not doing that!

I'm on the right track but it'll be a bit more work.  We need to use
sub-interpreters and thread switching; not sure how that'll impact
performance, but frankly we might be able to get around the GIL if we're
careful.