FINALLY.  Fixed a major leak.
I was creating new objects every time the script was invoked but not
cleaning up the old ones.  I'll need to put in some logic to more
gracefully handle changes and kill the old objects when necessary, as
I'm not currently doing that (so there's still a runtime leak there).

However, it's not an ever-increasing leak.  I need to properly destroy
returnList when I'm done with it, so I need to assess how to be done
with it (calling all the Python finalize methods).  I also need to
determine whether the input arguments to the weights_multi function is
the same or not, as it currently just spits back out the same objects it
created without bothering to see if the user wants a different set of
arguments.

It would be best if I properly destroyed the numpy objects and was able,
at least, to return fresh clean ones, but I can't be sure that
destroying the numpy objects doesn't also mark the Chapel array as being
"free".  And if Python is able to free to the memory that Chapel has
allocated, then that's a problem and I should more carefully invoke the
destruction method.