ONCE MORE WITH FEELING, FINALLY!
Finally.  Just finally.  Elliot helpfully suggested I use
AddressSanitizer, and it almost _immediately_ found the problem.  That
is, most Python macros return one of two things: what you expect them
to, or -1, for failure.

What a good person does is religiously check all user submitted input to
make sure it is what they think it is before making assumptions about
the fact that it's a tuple or a list.  In this case, I was counting the
total number of elements across all tuples and adding that up to
mallocate some memory.  However, Python internally treats some things
that _look_ like tuples as ints (basically, no such thing as a length 1
tuple), and so checking the length on these returns -1.  Which subtracts
from the length you need.  Which means you don't allocate enough memory
and corrupt everything.

I was also casting an unsigned long long array pointer to a double, then
freeing the double.  Yeah, that didn't work either.  That was resulting
in heap fragmentation.  And good luck getting them to reproduce cleanly.