Switched backends to theano for the moment.
Turns out Tensorflow doesn't play nicely with threads, and it wants to
control pretty much everything.  It looks like it releases the GIL,
which is a problem for us here.

That might mean just doing a TF interface manually, later.  For the
moment, it seems Theano doesn't release the GIL, so we can at least make
sure that's working appropriately.